% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotROC.r
\name{plotROC}
\alias{plotROC}
\title{plotROC() function}
\usage{
plotROC(prob, true, resolution = 1000)
}
\description{
Function to plot the ROC curve
}
\details{
input:
prob: vector/array/data.frame...: containing the predicted probabilities from 'predict' function
true: vector/array/data.frame...: indicating the true label; has to have the same length as prob
resolution: int: search resolution, default set to 1000


output:
A plot of the ROC curve
}
