% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROC.r
\name{ROC}
\alias{ROC}
\title{ROC() function}
\usage{
ROC(prob, true, resolution = 1000)
}
\description{
Function to get a data frame containing the ROC
}
\details{
input:
prob: vector/array/data.frame...: containing the predicted probabilities from 'predict' function
true: vector/array/data.frame...: indicating the true label; has to have the same length as prob
resolution: int: search resolution, default set to 1000


output:
The function returns an R data frame, each row containing the True Positive Rate (tpr)
and False Positive Rate (fpr), according to the same threshold.
}
