% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold.r
\name{threshold}
\alias{threshold}
\title{threshold() function}
\usage{
threshold(prob, true, resolution = 1000)
}
\description{
function to search for threshold
}
\details{
input:
prob: vector/array/data.frame...: containing the predicted probabilities from 'predict' function
true: vector/array/data.frame...: indicating the true label; has to have the same length as prob
resolution: int: search resolution, default set to 1000


output:
An R data frame containing the three optimal thresholds, according to the three metrics:
Accuracy, Balanced Accuracy and F1 Score
}
